requirements.txt
Flask
requests
mysql-connector-python
pytz
tzlocal
RPi.bme280
w1thermsensor